refactor: 도메인 테스트를 위한 기존 테스트 패키지 분리
